requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.0
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
idna==3.7
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
